Update to 2019
Sankaku now returns error 500 if you use an user-agent like curl or wget. So you have to add a fake user-agent to the request. Moreover, there is a need of to replace "&amp;" with "&" (if the post shows an image "sample" instead of "original" there is a "&amp;" instead of "&" in the Sankaku's html source.